separated Metric component logic to different files
